move test with circular dep from ldp to ssi::vc
