+++ ssi-ldp/src/eip712.rs
@@ -1678,233 +1678,6 @@ mod tests {
         }
     }
 
-    #[async_std::test]
-    async fn verify_typed_data() {
-        let proof: Proof = serde_json::from_value(json!({
-          "verificationMethod": "did:example:aaaabbbb#issuerKey-1",
-          "created": "2021-07-09T19:47:41Z",
-          "proofPurpose": "assertionMethod",
-          "type": "EthereumEip712Signature2021",
-          "eip712": {
-            "types": {
-              "EIP712Domain": [
-                { "name": "name", "type": "string" },
-                { "name": "version", "type": "string" },
-                { "name": "chainId", "type": "uint256" },
-                { "name": "salt", "type": "bytes32" }
-              ],
-              "VerifiableCredential": [
-                { "name": "@context", "type": "string[]" },
-                { "name": "type", "type": "string[]" },
-                { "name": "id", "type": "string" },
-                { "name": "issuer", "type": "string" },
-                { "name": "issuanceDate", "type": "string" },
-                { "name": "credentialSubject", "type": "CredentialSubject" },
-                { "name": "credentialSchema", "type": "CredentialSchema" },
-                { "name": "proof", "type": "Proof" }
-              ],
-              "CredentialSchema": [
-                { "name": "id", "type": "string" },
-                { "name": "type", "type": "string" }
-              ],
-              "CredentialSubject": [
-                { "name": "type", "type": "string" },
-                { "name": "id", "type": "string" },
-                { "name": "name", "type": "string" },
-                { "name": "child", "type": "Person" }
-              ],
-              "Person": [
-                { "name": "type", "type": "string" },
-                { "name": "name", "type": "string" }
-              ],
-              "Proof": [
-                { "name": "verificationMethod", "type": "string" },
-                { "name": "created", "type": "string" },
-                { "name": "proofPurpose", "type": "string" },
-                { "name": "type", "type": "string" }
-              ]
-            },
-            "primaryType": "VerifiableCredential",
-            "domain": {
-              "name": "https://example.com",
-              "version": "2",
-              "chainId": 4,
-              "salt": "0x000000000000000000000000000000000000000000000000aaaabbbbccccdddd"
-            }
-          }
-        }))
-        .unwrap();
-        let vc: ssi::vc::Credential = serde_json::from_value(json!({
-          "@context": [
-            "https://www.w3.org/2018/credentials/v1",
-            "https://schema.org"
-          ],
-          "type": [
-            "VerifiableCredential"
-          ],
-          "id": "https://example.org/person/1234",
-          "issuer": "did:example:aaaabbbb",
-          "issuanceDate": "2010-01-01T19:23:24Z",
-          "credentialSubject": {
-            "type": "Person",
-            "id": "did:example:bbbbaaaa",
-            "name": "Vitalik",
-            "child": {
-              "type": "Person",
-              "name": "Ethereum"
-            }
-          },
-          "credentialSchema": {
-            "id": "https://example.com/schemas/v1",
-            "type": "Eip712SchemaValidator2021"
-          }
-        }))
-        .unwrap();
-        let typed_data = TypedData::from_document_and_options_json(&vc, &proof)
-            .await
-            .unwrap();
-        // https://w3c-ccg.github.io/ethereum-eip712-signature-2021-spec/#example-5
-        let expected_typed_data = json!({
-          "types": {
-            "EIP712Domain": [
-              { "name": "name", "type": "string" },
-              { "name": "version", "type": "string" },
-              { "name": "chainId", "type": "uint256" },
-              { "name": "salt", "type": "bytes32" }
-            ],
-            "VerifiableCredential": [
-              { "name": "@context", "type": "string[]" },
-              { "name": "type", "type": "string[]" },
-              { "name": "id", "type": "string" },
-              { "name": "issuer", "type": "string" },
-              { "name": "issuanceDate", "type": "string" },
-              { "name": "credentialSubject", "type": "CredentialSubject" },
-              { "name": "credentialSchema", "type": "CredentialSchema" },
-              { "name": "proof", "type": "Proof" }
-            ],
-            "CredentialSchema": [
-              { "name": "id", "type": "string" },
-              { "name": "type", "type": "string" }
-            ],
-            "CredentialSubject": [
-              { "name": "type", "type": "string" },
-              { "name": "id", "type": "string" },
-              { "name": "name", "type": "string" },
-              { "name": "child", "type": "Person" }
-            ],
-            "Person": [
-              { "name": "type", "type": "string" },
-              { "name": "name", "type": "string" }
-            ],
-            "Proof": [
-              { "name": "verificationMethod", "type": "string" },
-              { "name": "created", "type": "string" },
-              { "name": "proofPurpose", "type": "string" },
-              { "name": "type", "type": "string" }
-            ]
-          },
-          "domain": {
-            "name": "https://example.com",
-            "version": "2",
-            "chainId": 4,
-            "salt": "0x000000000000000000000000000000000000000000000000aaaabbbbccccdddd"
-          },
-          "primaryType": "VerifiableCredential",
-          "message": {
-            "@context": [
-              "https://www.w3.org/2018/credentials/v1",
-              "https://schema.org"
-            ],
-            "type": [
-              "VerifiableCredential"
-            ],
-            "id": "https://example.org/person/1234",
-            "issuer": "did:example:aaaabbbb",
-            "issuanceDate": "2010-01-01T19:23:24Z",
-            "credentialSubject": {
-              "type": "Person",
-              "id": "did:example:bbbbaaaa",
-              "name": "Vitalik",
-              "child": {
-                "type": "Person",
-                "name": "Ethereum"
-              }
-            },
-            "credentialSchema": {
-              "id": "https://example.com/schemas/v1",
-              "type": "Eip712SchemaValidator2021"
-            },
-            "proof": {
-              "verificationMethod": "did:example:aaaabbbb#issuerKey-1",
-              "created": "2021-07-09T19:47:41Z",
-              "proofPurpose": "assertionMethod",
-              "type": "EthereumEip712Signature2021"
-            }
-          }
-        });
-        assert_eq!(
-            serde_json::to_value(&typed_data).unwrap(),
-            expected_typed_data
-        );
-
-        let jwk: ssi_jwk::JWK = serde_json::from_value(json!({
-            "kty": "EC",
-            "crv": "secp256k1",
-            "x": "cmbYyDC6cbm807_OmFNYP4CLEL0aB2F1UG683SxFkXM",
-            "y": "zBw5HAh0cJM4YimSQvtYM1HFhzUXVUgrDhxJ70aajt0",
-            "d": "u7QuEl6W0XNppEY0iMVjATT99tC9acwV3Z2keEqvKGo"
-        }))
-        .unwrap();
-        eprintln!("jwk {}", serde_json::to_string(&jwk).unwrap());
-
-        let td_jcs = serde_jcs::to_string(&typed_data).unwrap();
-        // Wrap string with line breaks
-        // https://stackoverflow.com/a/57032118
-        let jcs_lines = td_jcs
-            .chars()
-            .enumerate()
-            .flat_map(|(i, c)| {
-                if i != 0 && i % 90 == 0 {
-                    Some('\n')
-                } else {
-                    None
-                }
-                .into_iter()
-                .chain(std::iter::once(c))
-            })
-            .collect::<String>();
-        eprintln!("JCS: [\n{}\n]", jcs_lines);
-
-        // Sign proof
-        let bytes = typed_data.bytes().unwrap();
-        let ec_params = match &jwk.params {
-            ssi_jwk::Params::EC(ec) => ec,
-            _ => unreachable!(),
-        };
-        use k256::ecdsa::signature::Signer;
-        let secret_key = k256::SecretKey::try_from(ec_params).unwrap();
-        let signing_key = k256::ecdsa::SigningKey::from(secret_key);
-        let sig: k256::ecdsa::recoverable::Signature = signing_key.try_sign(&bytes).unwrap();
-        let sig_bytes = &mut sig.as_ref().to_vec();
-        // Recovery ID starts at 27 instead of 0.
-        sig_bytes[64] += 27;
-        let sig_hex = ssi_crypto::hashes::keccak::bytes_to_lowerhex(sig_bytes);
-        let mut proof = proof.clone();
-        proof.proof_value = Some(sig_hex.clone());
-        eprintln!("proof {}", serde_json::to_string(&proof).unwrap());
-
-        // Verify the VC/proof
-        let mut vc = vc.clone();
-        let mut context_loader = ssi_json_ld::ContextLoader::default();
-        vc.add_proof(proof.clone());
-        vc.validate().unwrap();
-        let verification_result = vc.verify(None, &DIDExample, &mut context_loader).await;
-        println!("{:#?}", verification_result);
-        assert!(verification_result.errors.is_empty());
-
-        assert_eq!(sig_hex, "0x5fb8f18f21f54c2df8a2720d0afcee7dbbb18e4b7a22ce6e8183633d63b076d329122584db769cd78b6cd5a7094ede5ceaa43317907539187f1f0d8875f99e051b");
-    }
-
     use crate::LinkedDataProofOptions;
     use ssi_core::uri::URI;
     use ssi_dids::VerificationRelationship as ProofPurpose;
